[restructure] move url related utils to ib/utils/url
